move utils for grabbing the namespace into a separate file
